+++ Magic/src/main/java/com/elmakers/mine/bukkit/wand/WandProperties.java
@@ -23,7 +23,7 @@ public abstract class WandProperties extends TemplatedProperties {
             "active_spell", "active_brush", "worn",
             "alternate_spell", "alternate_spell2", "alternate_spell3", "alternate_spell4",
             "alternate_spell5", "alternate_spell6", "alternate_spell7",
-            "bound", "has_uses", "uses", "upgrade", "indestructible", "use_mode",
+            "bound", "has_uses", "uses", "upgrade", "indestructible", "use_mode", "bound_name",
             "undroppable", "boostable",
             "effect_bubbles", "effect_color",
             "effect_particle", "effect_particle_count", "effect_particle_data",